config: update AC_PREREQ statement to 2.60
Unrelated to the previous patches, the new value simply reflects
the reality that the minimum level for autoconf to configure
all x.org modules in one pass is 2.60 dated June 2006.

A version later than 2.60 can be used, but no new features from such
a later version can be used in configure.ac.

Although the code in this module can be configured with a version earlier
than 2.60, most of code is now contained in macros from util-macros
which use features of version 2.60, at the present or in the future.

ftp://ftp.gnu.org/gnu/autoconf/autoconf-2.60.tar.gz

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>